requirements.txt
altgraph==0.17.2
auto-py-to-exe==2.10.1
bottle==0.12.19
bottle-websocket==0.2.9
cffi==1.14.6
click==8.0.1
colorama==0.4.4
DAWG-Python==0.7.2
docopt==0.6.2
Eel==0.12.4
future==0.18.2
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.2
inexactsearch==1.0.2
joblib==1.0.1
navec==0.10.0
nltk==3.6.3
numpy==1.21.2
pandas==1.3.3
pefile==2021.9.3
pycparser==2.20
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyspellchecker==0.6.2
python-dateutil==2.8.2
pytz==2021.3
pywin32-ctypes==0.2.0
regex==2021.9.30
scipy==1.7.1
silpa-common==0.3
six==1.16.0
soundex==1.1.3
tqdm==4.62.3
whichcraft==0.6.1
Whoosh==2.4.0
zope.event==4.5.0
zope.interface==5.4.0
